Update parsing of serial data (AT commands from user)
Current parsing was expecting full strings to appear in the serial buffer
at once ; while when connected to an USB-Serial adapter the entry
is more character per character.
This patch tries to fill the buffer was as many as characters as possible,
and process the line once a CR (\n) has been received - while trying to
handle buffer overflows.